--study

+++ src/main/java/com/mihadev/zebra/repository/ClassRepository.java
@@ -3,5 +3,5 @@ package com.mihadev.zebra.repository;
 import com.mihadev.zebra.entity.Clazz;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ClassRepository extends CrudRepository<Clazz, Long> {
+public interface ClassRepository extends CrudRepository<Clazz, Integer> {
 }
